+++ Source/CesiumRuntime/Public/CesiumUrlTemplateRasterOverlay.h
@@ -220,6 +220,12 @@ public:
       meta = (ClampMin = 64, ClampMax = 2048))
   int32 TileHeight = 256;
 
+  /**
+   * HTTP headers to be attached to each request made for this raster overlay.
+   */
+  UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Cesium")
+  TMap<FString, FString> RequestHeaders;
+
 protected:
   virtual std::unique_ptr<CesiumRasterOverlays::RasterOverlay> CreateOverlay(
       const CesiumRasterOverlays::RasterOverlayOptions& options = {}) override;